Add prompt folder
