+++ app/commands/enquire.js
@@ -16,7 +16,7 @@ import { generateCompletion, parseEnquiry } from '../../utils/index.js';
 import MessageAction from '../actions/message.js';
 import Context from '../context.js';
 import { getHistory, updateHistory } from '../histories.js';
-import { getPrompt, SENTENCE_ENQUIRING, setPrompt } from '../prompts.js';
+import { getPrompt, SENTENCE_ENQUIRING, setPrompt } from '../prompt/index.js';
 import { isCallCommand } from './call.js';
 
 const hasCommand = (context) => (command) => context.isCommand(command) || (isCallCommand(context) && context.hasCommand(command));
