+++ app/prompt/prompt.js
@@ -1,0 +1,45 @@
+import { t } from '../../languages/index.js';
+import { PARTICIPANT_AI } from '../../services/openai.js';
+import Sentence, { SENTENCE_PROMPTING } from './sentence.js';
+
+const MAX_LINE_COUNT = 16;
+
+class Prompt {
+  sentences = [];
+
+  constructor() {
+    this.write(PARTICIPANT_AI, t('__COMPLETION_INIT_MESSAGE'));
+  }
+
+  /**
+   * @returns {Sentence}
+   */
+  get lastSentence() {
+    return this.sentences.length > 0 ? this.sentences[this.sentences.length - 1] : null;
+  }
+
+  /**
+   * @param {string} title
+   * @param {string} text
+   */
+  write(title, text = '') {
+    if (this.sentences.length >= MAX_LINE_COUNT) {
+      this.sentences.shift();
+    }
+    this.sentences.push(new Sentence({ type: SENTENCE_PROMPTING, title, text }));
+    return this;
+  }
+
+  /**
+   * @param {string} text
+   */
+  patch(text) {
+    this.sentences[this.sentences.length - 1].text += text;
+  }
+
+  toString() {
+    return this.sentences.map((sentence) => sentence.toString()).join('');
+  }
+}
+
+export default Prompt;
